feat: Split up SavedItemFormView
